+++ papiea-engine/src/validator/index.ts
@@ -6,7 +6,7 @@ import {
     FieldBehavior,
     IntentfulBehaviour,
     Kind,
-    Metadata,
+    Metadata, Procedural_Signature,
     Provider,
     Spec,
     Status,
@@ -21,6 +21,7 @@ import { resolve } from "path"
 import { cloneDeep } from "lodash"
 import { PapieaException } from "../errors/papiea_exception"
 import uuid = require("uuid");
+import { Logger, LoggerFactory } from "papiea-backend-utils";
 
 // We can receive model in 2 forms:
 // As user specified in definition, which means it has "properties" field ( { properties: {} } } )
@@ -164,14 +165,15 @@ export class ValidatorImpl {
      * @param schema - schema to remove the fields from.
      * @param fieldName - type of x-papiea value spec-only|status-only.
      */
-    remove_schema_fields(schema: any, fieldName: string) {
-        for (let prop in schema) {
-            if (typeof schema[prop] === 'object' && "x-papiea" in schema[prop] && schema[prop]["x-papiea"] === fieldName) {
-                delete schema[prop]
-            } else if (typeof schema[prop] === 'object')
-                this.remove_schema_fields(schema[prop], fieldName)
-        }
-    }
+   remove_schema_fields(schema: any, fieldName: string) {
+       for (let prop in schema) {
+           if (typeof schema[prop] === "object" && "x-papiea" in schema[prop] && schema[prop]["x-papiea"] === fieldName) {
+               delete schema[prop]
+           } else if (typeof schema[prop] === "object") {
+               this.remove_schema_fields(schema[prop], fieldName)
+           }
+       }
+   }
 
     public async validate_status(provider: Provider, entity_ref: Entity_Reference, status: Status) {
         if (status === undefined || isEmpty(status)) {
@@ -201,18 +203,20 @@ export class ValidatorImpl {
         Object.assign(schemas, this.provider_schema)
         Object.assign(schemas, this.procedural_signature_schema)
         this.validate(
-            provider.prefix, provider.version, 'Provider',
+            provider.prefix, provider.version, "Provider",
             provider, Object.values(this.provider_schema)[0],
             schemas, true, Object.keys(this.provider_schema)[0], this.validate_provider.name, true)
         Object.values(provider.procedures).forEach(proc => {
+            this.check_nullable_modifier_procedure(proc, provider.prefix, provider.version)
             this.validate(
-                provider.prefix, provider.version, 'ProviderProcedure',
+                provider.prefix, provider.version, "ProviderProcedure",
                 proc, Object.values(this.procedural_signature_schema)[0],
                 schemas, true, Object.keys(this.procedural_signature_schema)[0],
                 proc.name, true)
         })
         provider.kinds.forEach(kind => {
             Object.values(kind.kind_procedures).forEach(proc => {
+                this.check_nullable_modifier_procedure(proc, provider.prefix, provider.version, kind.name)
                 this.validate(
                     provider.prefix, provider.version, kind.name,
                     proc, Object.values(this.procedural_signature_schema)[0],
@@ -220,6 +224,7 @@ export class ValidatorImpl {
                     proc.name, true)
             })
             Object.values(kind.entity_procedures).forEach(proc => {
+                this.check_nullable_modifier_procedure(proc, provider.prefix, provider.version, kind.name)
                 this.validate(
                     provider.prefix, provider.version, kind.name,
                     proc, Object.values(this.procedural_signature_schema)[0],
@@ -227,6 +232,7 @@ export class ValidatorImpl {
                     proc.name, true)
             })
             Object.values(kind.intentful_signatures).forEach(proc => {
+                this.check_nullable_modifier_procedure(proc, provider.prefix, provider.version, kind.name)
                 this.validate(
                     provider.prefix, provider.version, kind.name,
                     proc, Object.values(this.procedural_signature_schema)[0],
@@ -242,11 +248,15 @@ export class ValidatorImpl {
     validate_kind_structure(schema: Data_Description, provider_prefix: string, provider_version: string, kind_name: string) {
         const x_papiea_field = "x-papiea"
         const status_only_value = FieldBehavior.StatusOnly
+        this.check_nullable_modifier(schema, provider_prefix, provider_version, kind_name)
         // x_papiea_field property have only status_only_value value
         this.validate_field_value(schema[kind_name], x_papiea_field, [status_only_value], provider_prefix, provider_version, kind_name)
         this.validate_spec_only_structure(schema[kind_name], provider_prefix, provider_version, kind_name)
         // status-only fields cannot be required in schema
         this.validate_status_only_field(schema, provider_prefix, provider_version, kind_name)
+        // warn for untyped object which are not marked as status-only in schema
+        const logger = LoggerFactory.makeLogger({});
+        this.validate_untyped_object(schema, provider_prefix, provider_version, kind_name, logger)
     }
 
     validate_field_value(schema: Data_Description, field_name: string, possible_values: string[], provider_prefix: string, provider_version: string, kind_name: string) {
@@ -283,6 +293,32 @@ export class ValidatorImpl {
         }
     }
 
+    check_nullable_modifier(schema: Data_Description, provider_prefix: string, provider_version: string, kind_name?: string) {
+        for (let field in schema) {
+            if (!schema.hasOwnProperty(field)) {
+                continue
+            }
+            const field_schema = schema[field]
+            if (field_schema.hasOwnProperty("type")) {
+                if (field_schema["type"] === "object") {
+                    this.check_nullable_modifier(field_schema["properties"], provider_prefix, provider_version, kind_name)
+                }
+                if (field_schema.hasOwnProperty("nullable")) {
+                    const message = `Papiea doesn't support 'nullable' fields. Please make a field '${field}' non-required instead. for: ${provider_prefix}/${provider_version}`
+                    throw new ValidationError({
+                        message: kind_name ? message + `/${kind_name}` : message,
+                        entity_info: { provider_prefix: provider_prefix, provider_version: provider_version, kind_name: kind_name }
+                    })
+                }
+            }
+        }
+    }
+
+    check_nullable_modifier_procedure(proc: Procedural_Signature, provider_prefix: string, provider_version: string, kind_name?: string) {
+       this.check_nullable_modifier(proc.argument, provider_prefix, provider_version, kind_name)
+       this.check_nullable_modifier(proc.result, provider_prefix, provider_version, kind_name)
+    }
+
     validate_status_only_field(schema: Data_Description, provider_prefix: string, provider_version: string, kind_name: string) {
         try {
             for(let field in schema) {
@@ -314,6 +350,34 @@ export class ValidatorImpl {
                                 }
                                 this.validate_status_only_field(field_schema["items"]["properties"], provider_prefix, provider_version, kind_name)
                             }
+                        }
+                    }
+                }
+            }
+        } catch (e) {
+            throw (e)
+        }
+    }
+
+    validate_untyped_object(schema: Data_Description, provider_prefix: string, provider_version: string, kind_name: string, logger: Logger, field_name: string = '') {
+        try {
+            for(let field in schema) {
+                const field_schema = schema[field]
+                if (field_schema.hasOwnProperty("type")) {
+                    if (field_schema["type"] === "object") {
+                        if ((!field_schema.hasOwnProperty("properties") || field_schema["properties"].length === 0) && (!field_schema.hasOwnProperty("x-papiea") || field_schema["x-papiea"] !== "status-only")) {
+                            logger.warn(`Field ${field_name + "/" + field} is an untyped object with no properties for kind: ${provider_prefix}/${provider_version}/${kind_name}. Please check if this is expected.`)
+                            return
+                        }
+                        this.validate_untyped_object(field_schema["properties"], provider_prefix, provider_version, kind_name, logger, field_name + "/" + field)
+                    }
+                    if (field_schema["type"] === "array") {
+                        if (field_schema.hasOwnProperty("items") && field_schema["items"].hasOwnProperty("type") && field_schema["items"]["type"].includes("object")) {
+                            if ((!field_schema.hasOwnProperty("properties") || field_schema["properties"].length === 0) && (!field_schema.hasOwnProperty("x-papiea") || field_schema["x-papiea"] !== "status-only")) {
+                                logger.warn(`Field ${field_name + "/" + field} is an untyped object with no properties for kind: ${provider_prefix}/${provider_version}/${kind_name}. Please check if this is expected.`)
+                                return
+                            }
+                            this.validate_untyped_object(field_schema["items"]["properties"], provider_prefix, provider_version, kind_name, logger, field_name + "/" + field)
                         }
                     }
                 }